refactor(base): split connect_loop.rs by handler group (#244)
connect_loop.rs was 730 lines (over the 700-line cap), driven by the
giant inbound-msg-id match in handle_encrypted_datagram. Split by
handler group as the issue suggested:

  connect_loop/mod.rs         — run_connect_loop, handle_datagram,
                                CONSTANT/WORD payload scanners, tests
  connect_loop/encrypted.rs   — handle_encrypted_datagram (decrypt +
                                bundle scan + match shell)
  connect_loop/account_arms.rs — 0xC2..=0xC7 base-method dispatch
                                 (account select vs. SGWPlayer in-world)
  connect_loop/cell_arms.rs   — 0x80..=0xBF cell-method dispatch
                                (direct + 0xBD sub-slot encoding)

Public surface preserved: parent declares
`pub(crate) mod connect_loop;` and only `run_connect_loop` is referenced
externally; mod.rs re-exports `handle_encrypted_datagram` from the new
encrypted module so anyone reaching the crate-private path keeps working.

Cell-method dispatch returns `ControlFlow` so the original `continue`
inside the bundle-scan loop is preserved exactly.

All 3 scanner tests pass; clippy clean; cargo fmt applied.
